### 2024-06-10 #6865 @editorialbot
[PRE REVIEW]: anywidget: reusable widgets for interactive analysis and visualization in computational notebooks
**Submitting author:** @manzt (<a href="http://orcid.org/0000-0001-7694-5164">Trevor Manz</a>)
**Repository:** https://github.com/manzt/anywidget
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.9.12
