### 2024-06-10 #6865 @danielskatz
@manzt - Thanks for your submission.  While I look for an editor, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, if you have entries in your bibtex file that you aren't using, they are still being checked, so rather than fixing them, you could remove them from your file.
