### 2024-06-10 53748c2
chore(paper): Add SciPy DOI
