+++ paper/paper.bib
@@ -577,11 +577,14 @@
   copyright = "http://creativecommons.org/licenses/by/4.0/"
 }
 
-@unpublished{scipy,
-  author = {Trevor Manz and Nils Gehlenborg and Nezar Abdennur},
-  title = {Any notebook served: authoring and sharing reusable interactive widgets},
-  note = {Unpublished, submitted to Scipy 2024},
-  year = {2024}
+@misc{scipy,
+ title={Any notebook served: authoring and sharing reusable interactive widgets},
+ url={osf.io/pyn7u},
+ DOI={10.31219/osf.io/pyn7u},
+ publisher={OSF Preprints},
+ author={Manz, Trevor and Gehlenborg, Nils and Abdennur, Nezar},
+ year={2024},
+ month={Jun}
 }
 
 @misc{ipywidgets,
