### 2024-06-11 #6865 @manzt
May I request the branch for this submission is switched to [`joss`](https://github.com/openjournals/joss-reviews/issues/6865#issue-2344714861) (rather than main)?
