+++ paper/paper.bib
@@ -56,7 +56,8 @@
   year      =  2023,
   address   = "New York, NY, USA",
   keywords  = "computational notebooks, code search, software visualization,
-               Exploratory data analysis"
+               Exploratory data analysis",
+  doi={10.1145/3545995}
 }
 
 % The entry below contains non-ASCII chars that could not be converted
@@ -108,7 +109,8 @@
   month     =  sep,
   year      =  2022,
   copyright = "http://creativecommons.org/licenses/by-nc/4.0/",
-  language  = "en"
+  language  = "en",
+  doi={10.1002/spe.3098}
 }
 
 @INPROCEEDINGS{Wang2023,
@@ -146,8 +148,7 @@
   address   = "New York, NY, USA",
   keywords  = "computational notebooks, data science., human-AI collaboration,
                natural language processing, slides generation",
-  location  = "<conf-loc>, <city>Hamburg</city>, <country>Germany</country>,
-               </conf-loc>"
+  doi={10.1145/3544548.3580753}
 }
 
 @INPROCEEDINGS{Drosos2020,
@@ -182,8 +183,7 @@
   year      =  2020,
   address   = "New York, NY, USA",
   keywords  = "computational notebooks, data science, program synthesis",
-  location  = "<conf-loc>, <city>Honolulu</city>, <state>HI</state>,
-               <country>USA</country>, </conf-loc>"
+  doi={10.1145/3313831.3376442}
 }
 
 @INPROCEEDINGS{Li2023,
@@ -218,8 +218,7 @@
   year      =  2023,
   address   = "New York, NY, USA",
   keywords  = "computational notebooks, data storytelling, data visualization",
-  location  = "<conf-loc>, <city>Hamburg</city>, <country>Germany</country>,
-               </conf-loc>"
+  doi={10.1145/3544548.3580965}
 }
 
 @INPROCEEDINGS{Jain2022,
@@ -256,7 +255,8 @@
   month     =  jul,
   year      =  2022,
   address   = "New York, NY, USA",
-  location  = "Pittsburgh, Pennsylvania"
+  location  = "Pittsburgh, Pennsylvania",
+  doi={10.1145/3510003.3510203}
 }
 
 @ARTICLE{Wang2022,
@@ -287,7 +287,8 @@
   year          =  2022,
   archivePrefix = "arXiv",
   primaryClass  = "cs.HC",
-  eprint        = "2205.03963"
+  eprint        = "2205.03963",
+  doi={10.48550/arXiv.2205.03963}
 }
 
 @INPROCEEDINGS{Wang2024,
@@ -325,8 +326,7 @@
   address   = "New York, NY, USA",
   keywords  = "Computational Notebook, Cross-Platform Visualization, Data
                Science, Design, Interactive Visualization, Systematic Review",
-  location  = "<conf-loc> <city>Honolulu</city> <state>HI</state>
-               <country>USA</country> </conf-loc>"
+  doi={10.1145/3613905.3650848}
 }
 
 @INPROCEEDINGS{Kluyver2016,
@@ -352,7 +352,8 @@
   year       =  2016,
   language   = "en",
   conference = "20th International Conference on Electronic Publishing
-                (01/01/16)"
+                (01/01/16)",
+  doi={10.3233/978-1-61499-649-1-87}
 }
 
 @ARTICLE{Perez2007,
@@ -369,7 +370,8 @@
   volume    =  9,
   number    =  3,
   pages     = "21--29",
-  year      =  2007
+  year      =  2007,
+  doi={10.1109/MCSE.2007.53}
 }
 
[…]
