### 2024-06-11 0979976
Update CITATION.cff
