+++ CITATION.cff
@@ -1,32 +1,11 @@
-# This CITATION.cff file was generated with cffinit.
-# Visit https://bit.ly/cffinit to generate yours today!
-
 cff-version: 1.2.0
-title: anywidget
-message: >-
-  If you use this software, please cite it using the
-  metadata from this file.
-type: software
+title: "anywidget"
 authors:
-  - given-names: Trevor
-    family-names: Manz
-    email: trevor.j.manz@gmail.com
-    affiliation: Harvard Medical School
+  - given-names: "Trevor"
+    family-names: "Manz"
     orcid: "https://orcid.org/0000-0001-7694-5164"
-identifiers:
-  - type: doi
-    value: 10.5281/zenodo.8010549
-    description: Zenodo
-repository-code: "https://github.com/manzt/anywidget"
-url: "https://anywidget.dev"
-abstract: >-
-  anywidget is a Python library that simplifies the creation
-  and maintenance of Jupyter Widgets, making them more
-  accessible for computational scientists. By leveraging
-  ECMAScript modules (ESM), it addresses complex widget
-  packaging issues, enhancing the utility and
-  interoperability of custom widgets across different
-  environments.
+    affiliation: "Harvard Medical School"
+url: "https://github.com/manzt/anywidget"
 keywords:
   - Jupyter Notebooks
   - Jupyter Widgets
@@ -34,3 +13,23 @@ keywords:
   - Interactive Computing
   - Data Analysis
 license: MIT
+preferred-citation:
+  type: article
+  title: "anywidget: reusable widgets for interactive analysis and visualization in computational notebooks"
+  authors:
+    - given-names: "Trevor"
+      family-names: "Manz"
+      orcid: "https://orcid.org/0000-0001-7694-5164"
+      affiliation: "Harvard Medical School"
+    - given-names: "Nezar"
+      family-names: "Abdennur"
+      affiliation: "University of Massachusetts Chan Medical School"
+      orcid: "https://orcid.org/0000-0001-5814-0864"
+    - given-names: "Nils"
+      family-names: "Gehlenborg"
+      affiliation: "Harvard Medical School"
+      orcid: "https://orcid.org/0000-0003-0327-8297"
+  doi: "10.31219/osf.io/tw9sg"
+  journal: "OSF preprint"
+  month: 6
+  year: 2024
