### 2024-06-11 #6865 @manzt
> While I look for an editor, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect.

I have added the available missing DOIs to the [`joss`](https://github.com/openjournals/joss-reviews/issues/6865#issue-2344714861) branch.
