### 2024-06-12 9d1f9e5
fix: Drop CJS support
