### 2024-06-12 dee4d6f
feat(vite): Support default export of render in Vite HMR wrapper (#598)
Co-authored-by: Quentin Ågren <quentin@scilifelab.se>
Co-authored-by: Trevor Manz <trevor.j.manz@gmail.com>
