+++ packages/vite/index.js
@@ -1,57 +1,15 @@
-/** @type {(src: string) => string} */
-let template = (src) => `
-function noop() {}
+import * as fs from "node:fs";
+import * as path from "node:path";
+import * as url from "node:url";
 
-function emptyElement(el) {
-	while (el.firstChild) {
-		el.removeChild(el.firstChild);
-	}
-}
+const query = "?anywidget";
+const namespace = "anywidget:";
+const resolvedNamespace = `\0${namespace}`;
 
-function showErrorOverlay(err) {
-	let ErrorOverlay = customElements.get("vite-error-overlay");
-	// don't open outside vite environment
-	if (!ErrorOverlay) return;
-	let overlay = new ErrorOverlay(err);
-	document.body.appendChild(overlay);
-}
-
-window.addEventListener("error", showErrorOverlay);
-window.addEventListener("unhandledrejection", (e) => showErrorOverlay(e.reason));
-
-import.meta.hot.accept("${src}", (newModule) => {
-	import.meta.hot.data.render = newModule.render;
-	refresh();
+let dirname = path.dirname(url.fileURLToPath(import.meta.url));
+let hmrTemplate = fs.readFileSync(path.resolve(dirname, "hmr.js"), {
+	encoding: "utf-8",
 });
-
-export async function render({ model, el } ) {
-	if (import.meta.hot.data.render == null) {
-		let m = await import("${src}");
-		import.meta.hot.data.render = m.render;
-	}
-	if (import.meta.hot.data.contexts == null) {
-		import.meta.hot.data.contexts = [];
-	}
-	import.meta.hot.data.contexts.push({ cleanup: noop, model, el });
-	refresh();
-}
-
-async function refresh() {
-	let render = import.meta.hot.data.render;
-	for (let context of import.meta.hot.data.contexts) {
-		try {
-			await context.cleanup();
-		} catch (e) {
-			console.debug("[anywidget] error cleaning up previous module.", e);
-			context.cleanup = noop;
-		}
-		context.model.off();
-		emptyElement(context.el);
-		let cleanup = await render({ model: context.model, el: context.el });
-		context.cleanup = cleanup ?? noop;
-	}
-}
-`;
 
 /** @returns {import("vite").Plugin} */
 export default function () {
@@ -59,21 +17,22 @@ export default function () {
 		name: "anywidget",
 		apply: "serve",
 		resolveId(id) {
-			if (id.startsWith("anywidget:")) {
+			if (id.startsWith(namespace)) {
 				return `\0${id}`;
 			}
 		},
-		load(id) {
-			if (id.startsWith("\0anywidget:")) {
-				return template(id.split(":")[1]);
+		async load(id) {
+			if (id.startsWith(resolvedNamespace)) {
+				let src = id.split(":")[1];
+				return hmrTemplate.replaceAll("__ANYWIDGET_HMR_SRC__", src);
 			}
 		},
 		configureServer(server) {
 			server.middlewares.use((req, _res, next) => {
-				if (req.url?.endsWith("?anywidget")) {
+				if (req.url?.endsWith(query)) {
 					// turn into a bare identifier
-					let path = req.url.slice(0, -"?anywidget".length);
-					req.url = `anywidget:${path}`;
+					let path = req.url.slice(0, -query.length);
+					req.url = `${namespace}${path}`;
 				}
 				next();
 			});
