### 2024-06-12 0f414f3
ci: Remove JOSS GH Action from main
