+++ .changeset/breezy-wolves-grow.md
@@ -1,5 +1,0 @@
----
-"@anywidget/vite": minor
----
-
-Support anywidget lifecycle APIs (i.e., `default` export) in Vite plugin. Drops support for CJS-based usage of the plugin. The CJS Node API of Vite is deprecated in v5. Follow the [migration guide](https://vitejs.dev/guide/migration#deprecate-cjs-node-api) for more details.
