+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @anywidget/vite
 
+## 0.2.0
+
+### Minor Changes
+
+- Support anywidget lifecycle APIs (i.e., `default` export) in Vite plugin. Drops support for CJS-based usage of the plugin. The CJS Node API of Vite is deprecated in v5. Follow the [migration guide](https://vitejs.dev/guide/migration#deprecate-cjs-node-api) for more details. ([#598](https://github.com/manzt/anywidget/pull/598))
+
 ## 0.1.2
 
 ### Patch Changes
