### 2024-06-19 #6865 @hugoledoux
@manzt I am looking for reviewers, but while reading the paper, I noticed that you put an abstract: this is unnecessary since JOSS papers are concise. So if you could at some point remove it it would be good.
