### 2024-06-19 6742aee
chore: Remove abstract
