+++ paper/paper.md
@@ -1,15 +1,5 @@
 ---
 title: 'anywidget: reusable widgets for interactive analysis and visualization in computational notebooks'
-abstract: |
-  The anywidget project provides a specification and toolset for creating
-  portable web-based widgets for computational notebooks. It defines a standard
-  for front-end widget code using ECMAScript modules and offers tools for
-  authoring, distributing, and executing these modules across Jupyter-compatible
-  platforms (e.g., JupyterLab, Google Colab, VS Code). The project has spurred
-  the creation of many interactive visualization libraries and has been
-  integrated into existing ones. The standard has been embraced by other
-  notebook platforms and web frameworks, extending the reach of interactive
-  widgets beyond the Jupyter ecosystem.
 tags:
   - Computational notebooks
   - Jupyter
