### 2024-06-21 670967e
chore: Improve deprecation notice information (#609)
