+++ .changeset/smart-insects-join.md
@@ -1,0 +1,5 @@
+---
+"anywidget": patch
+---
+
+Improve legacy ESM deprecation notice with widget provenance and Python migration instruction
