+++ .changeset/smart-insects-join.md
@@ -1,5 +1,0 @@
----
-"anywidget": patch
----
-
-Improve legacy ESM deprecation notice with widget provenance and Python migration instruction
