+++ .changeset/silent-kids-shop.md
@@ -1,0 +1,5 @@
+---
+"create-anywidget": minor
+---
+
+Upgrade templates to TypeScript v5.5
