+++ .changeset/silent-kids-shop.md
@@ -1,5 +1,0 @@
----
-"create-anywidget": minor
----
-
-Upgrade templates to TypeScript v5.5
