### 2024-06-26 f8189bb
feat: Implement hoisting
