+++ packages/anywidget/src/widget.js
@@ -244,6 +244,72 @@ function throw_anywidget_error(source) {
 }
 
 /**
+ * @param {unknown} v
+ * @return {v is {}}
+ */
+function is_object(v) {
+	return typeof v === "object" && v !== null;
+}
+
+/**
+ * @param {unknown} v
+ * @return {v is import("@jupyter-widgets/base").DOMWidgetModel}
+ */
+function is_model(v) {
+	return is_object(v) && "on" in v && typeof v.on === "function";
+}
+
+/**
+ * @template {"_esm" | "_css"} T
+ * @param {import("@jupyter-widgets/base").DOMWidgetModel} model
+ * @param {T} asset_name
+ * @returns {{ get(name: T): string, on(event: `change:${T}`, callback: () => void): void, off(event: `change:${T}`): void }}
+ */
+function resolve_asset_model(model, asset_name) {
+	let value = model.get(asset_name);
+	if (is_model(value)) {
+		return {
+			/** @param {T} _name */
+			get(_name) {
+				return value.get("data");
+			},
+			/**
+			 * @param {`change:${T}`} _event
+			 * @param {() => void} callback
+			 */
+			on(_event, callback) {
+				value.on("change:data", callback);
+			},
+			/**
+			 * @param {`change:${T}`} _event
+			 */
+			off(_event) {
+				return value.off("change:data");
+			},
+		};
+	}
+	return model;
+}
+
+/**
+ * @template {"_esm" | "_css"} T
+ * @param {import("@jupyter-widgets/base").DOMWidgetModel} base_model
+ * @param {T} asset_name
+ * @param {() => void} cb
+ */
+function create_asset_signal(base_model, asset_name, cb) {
+	let model = resolve_asset_model(base_model, asset_name);
+	/** @type {import("solid-js").Signal<string>} */
+	let [asset, set_asset] = solid.createSignal(model.get(asset_name));
+	model.on(`change:${asset_name}`, () => {
+		cb();
+		set_asset(model.get(asset_name));
+	});
+	solid.onCleanup(() => model.off(`change:${asset_name}`));
+	return asset;
+}
+
+/**
  * @typedef InvokeOptions
  * @prop {DataView[]} [buffers]
  * @prop {AbortSignal} [signal]
@@ -301,25 +367,18 @@ class Runtime {
 
 	/** @param {import("@jupyter-widgets/base").DOMWidgetModel} model */
 	constructor(model) {
+		let id = () => model.get("_anywidget_id");
+
 		this.#disposer = solid.createRoot((dispose) => {
-			let [css, set_css] = solid.createSignal(model.get("_css"));
-			model.on("change:_css", () => {
-				let id = model.get("_anywidget_id");
-				console.debug(`[anywidget] css hot updated: ${id}`);
-				set_css(model.get("_css"));
+			let css = create_asset_signal(model, "_css", () => {
+				console.debug(`[anywidget] css hot updated: ${id()}`);
 			});
-			solid.createEffect(() => {
-				let id = model.get("_anywidget_id");
-				load_css(css(), id);
+			solid.createEffect(() => load_css(css(), id()));
+
+			let esm = create_asset_signal(model, "_esm", () => {
+				console.debug(`[anywidget] esm hot updated: ${id()}`);
 			});
 
-			/** @type {import("solid-js").Signal<string>} */
-			let [esm, setEsm] = solid.createSignal(model.get("_esm"));
-			model.on("change:_esm", async () => {
-				let id = model.get("_anywidget_id");
-				console.debug(`[anywidget] esm hot updated: ${id}`);
-				setEsm(model.get("_esm"));
-			});
 			/** @type {void | (() => import("vitest").Awaitable<void>)} */
 			let cleanup;
 			this.#widget_result = solid.createResource(esm, async (update) => {
