### 2024-06-28 #6939 @editorialbot
[REVIEW]: anywidget: reusable widgets for interactive analysis and visualization in computational notebooks
**Submitting author:** @manzt (<a href="http://orcid.org/0000-0001-7694-5164">Trevor Manz</a>)
**Repository:** https://github.com/manzt/anywidget
**Branch with paper.md** (empty if default branch): main
**Version:** create-anywidget@0.7.3
