### 2024-06-28 9d09d3a
feat: Add signals lib
