+++ packages/signals/npm.README.md
@@ -1,0 +1,8 @@
+# @anywidget/signals
+
+Use signals with [anywidget](https://anywidget.dev)
+
+This is a stub for npm. The actual package is published on the
+[JavaScript Registry](https://jsr.io).
+
+Please see [`jsr:@anywidget/signals`](https://jsr.io/@anywidget/signals).
