+++ packages/signals/mod.ts
@@ -1,3 +1,26 @@
+/**
+ * @module
+ *
+ * Thie module contains an anywidget bridge to define widgets using signals.
+ *
+ * @example
+ * ```ts
+ * import { effect, signal } from "@preact/signals-core";
+ * import { defineWidget } from "@anywidget/signals";
+ *
+ * export default signalify(signal, {
+ *   render({ model, el }) {
+ *     let btn = document.createElement("button");
+ *     btn.addEventListener("click", () => model.value += 1);
+ *     effect(() => {
+ *       btn.innerText = `Count is ${model.value}`;
+ *     });
+ *     el.appendChild(btn);
+ *   },
+ * });
+ * ```
+ */
+
 import type * as aw from "@anywidget/types";
 
 // https://github.com/tc39/proposal-signals
@@ -10,18 +33,25 @@ type CustomMessageListener = {
 	(event: MessageEvent<[data: unknown, buffers: Array<Uint8Array>]>): void;
 };
 
-interface HostPlatform {
+/** The interface for a host platform connection */
+export interface HostPlatform {
+	/** Send a message to the host platform */
 	postMessage(msg: unknown, buffers?: Array<Uint8Array>): void;
+	/** Add a listener for custom messages from the host */
 	addEventListener(type: "message", listener: CustomMessageListener): void;
+	/** Remove a listener for custom messages from the host */
 	removeEventListener(type: "message", listener: CustomMessageListener): void;
 }
 
-type SignalModel<T extends Record<string, unknown>> = {
-	[Key in Exclude<keyof T & string, "host">]: T[Key];
-} & {
-	host: HostPlatform;
-};
+export type SignalModel<T extends Record<string, unknown>> =
+	& {
+		[Key in Exclude<keyof T & string, "host">]: T[Key];
+	}
+	& {
+		host: HostPlatform;
+	};
 
+/** Connects a Jupyter Widget model as a host platform */
 class JupyterWidgetHostPlatform extends EventTarget {
 	#model: aw.AnyModel;
 	constructor(model: aw.AnyModel) {
@@ -30,13 +60,20 @@ class JupyterWidgetHostPlatform extends EventTarget {
 		model.on("msg:custom", this.#dispatch.bind(this));
 	}
 	#dispatch(msg: unknown, buffers?: Array<Uint8Array>) {
-		this.dispatchEvent(new MessageEvent("message", { data: [msg, buffers] }));
+		this.dispatchEvent(
+			new MessageEvent("message", { data: [msg, buffers] }),
+		);
 	}
 	postMessage(msg: unknown, buffers?: Array<Uint8Array>) {
 		this.#model.send(msg, {}, buffers);
 	}
 }
 
+/**
+ * Create a signal model from a model and a signal function
+ * @param signal The function to create a TC39-like signal
+ * @param model The anywidget model to connect to
+ */
 function create<T extends Record<string, unknown>>(
 	signal: <T>(value: T) => Signal<T>,
 	model: aw.AnyModel<T>,
@@ -77,20 +114,20 @@ function create<T extends Record<string, unknown>>(
 	});
 }
 
-type CreateSignalFunction =
-	| (<T>(value: T) => Signal<T>)
-	| (<T>(value: T) => { value: T })
-	| (<T>(value: T) => [() => T, (value: T) => void]);
+type AnySignal = <T>(value: T) => Signal<T>;
+type PreactLikeSignal = <T>(value: T) => { value: T };
+type SolidLikeSignal = <T>(value: T) => [() => T, (value: T) => void];
+type CreateSignalFunction = AnySignal | PreactLikeSignal | SolidLikeSignal;
 
 function isPreactLikeSignal(
[…]
