+++ packages/signals/mod.ts
@@ -1,0 +1,211 @@
+/**
+ * @module
+ *
+ * Thie module contains an anywidget bridge to define widgets using signals.
+ *
+ * @example
+ * ```ts
+ * import { effect, signal } from "@preact/signals-core";
+ * import { defineWidget } from "@anywidget/signals";
+ *
+ * export default signalify(signal, {
+ *   render({ model, el }) {
+ *     let btn = document.createElement("button");
+ *     btn.addEventListener("click", () => model.value += 1);
+ *     effect(() => {
+ *       btn.innerText = `Count is ${model.value}`;
+ *     });
+ *     el.appendChild(btn);
+ *   },
+ * });
+ * ```
+ */
+
+import type * as aw from "@anywidget/types";
+
+// https://github.com/tc39/proposal-signals
+interface Signal<T> {
+	get(): T;
+	set(value: T): void;
+}
+
+type CustomMessageListener = {
+	(event: MessageEvent<[data: unknown, buffers: Array<Uint8Array>]>): void;
+};
+
+/** The interface for a host platform connection */
+export interface HostPlatform {
+	/** Send a message to the host platform */
+	postMessage(msg: unknown, buffers?: Array<Uint8Array>): void;
+	/** Add a listener for custom messages from the host */
+	addEventListener(type: "message", listener: CustomMessageListener): void;
+	/** Remove a listener for custom messages from the host */
+	removeEventListener(type: "message", listener: CustomMessageListener): void;
+}
+
+export type SignalModel<T extends Record<string, unknown>> = {
+	[Key in Exclude<keyof T & string, "host">]: T[Key];
+} & {
+	host: HostPlatform;
+};
+
+/** Connects a Jupyter Widget model as a host platform */
+class JupyterWidgetHostPlatform extends EventTarget {
+	#model: aw.AnyModel;
+	constructor(model: aw.AnyModel) {
+		super();
+		this.#model = model;
+		model.on("msg:custom", this.#dispatch.bind(this));
+	}
+	#dispatch(msg: unknown, buffers?: Array<Uint8Array>) {
+		this.dispatchEvent(new MessageEvent("message", { data: [msg, buffers] }));
+	}
+	postMessage(msg: unknown, buffers?: Array<Uint8Array>) {
+		this.#model.send(msg, {}, buffers);
+	}
+}
+
+/**
+ * Create a signal model from a model and a signal function
+ * @param signal The function to create a TC39-like signal
+ * @param model The anywidget model to connect to
+ */
+function create<T extends Record<string, unknown>>(
+	signal: <T>(value: T) => Signal<T>,
+	model: aw.AnyModel<T>,
+): SignalModel<T> {
+	// TODO: This could be a lot more simple if we infer all the attributes from the model
+	// Instead we dynamically create signals when the are first accessed
+	// deno-lint-ignore no-explicit-any
+	let signalModel: SignalModel<T> = {} as any;
+	Object.defineProperty(signalModel, "host", {
+		value: new JupyterWidgetHostPlatform(model),
+		writable: false,
+		enumerable: false,
+		configurable: false,
+	});
+	// @ts-expect-error - We are dynamically creating signals
+	let cache: Record<keyof T, Signal<unknown>> = {};
+	function signalFor<K extends keyof T>(key: K): Signal<T[K]> {
+		if (!(key in cache)) {
+			cache[key] = signal(model.get(key));
+		}
+		// @ts-expect-error - We have the right type
+		return cache[key];
+	}
+	return new Proxy(signalModel, {
+		get(target, key, receiver) {
[…]
