+++ .github/workflows/jsr.yml
@@ -1,4 +1,4 @@
-name: Publish @anywidget/deno to jsr
+name: Publish to jsr
 
 on:
   push:
@@ -13,6 +13,11 @@ jobs:
   publish:
     if: github.repository == 'manzt/anywidget'
     runs-on: macos-14
+    strategy:
+      matrix:
+        directory:
+          - packages/deno
+          - packages/signals
     permissions:
       contents: read
       id-token: write
@@ -22,7 +27,7 @@ jobs:
         with:
           deno-version: v1.x
       - run: |
-          cd packages/deno
+          cd ${{ matrix.directory }}
           # sync deno version with package.json
           sed -i '' 's/\"version\": \"[^\"]*\"/\"version\": \"'$(jq -r '.version' package.json)'\"/' deno.json
           deno publish --allow-dirty
