### 2024-06-29 969bbfc
feat(signals): Update Symbol documentation
