### 2024-06-29 866f197
fix(signals): Typos in README.md
