### 2024-06-29 1cafccc
fix(signals): Ugh more typos. Rename signalify -> defineWidget
