+++ packages/signals/README.md
@@ -12,7 +12,7 @@ Use [signals](https://github.com/tc39/proposal-signals) with
 import { effect, signal } from "@preact/signals-core";
 import { defineWidget } from "@anywidget/signals";
 
-export default signalify(signal, {
+export default defineWidget(signal, {
   render({ model, el }) {
     let btn = document.createElement("button");
     btn.addEventListener("click", () => model.value += 1);
@@ -44,7 +44,7 @@ With `@preact/signals-core`:
 import { effect, signal } from "@preact/signals-core";
 import { defineWidget } from "@anywidget/signals";
 
-export default signalify(signal, {
+export default defineWidget(signal, {
   render({ model, el }) {
     let btn = document.createElement("button");
     btn.addEventListener("click", () => model.value += 1);
