### 2024-07-04 #6939 @kylebarron
Not sure of the best way to access the PDF version of the paper, so I'm trying that
