### 2024-07-12 dfc3a6e
chore(deps-dev): bump the dev-dependencies group across 1 directory with 8 updates (#620)
* chore(deps-dev): bump the dev-dependencies group across 1 directory with 8 updates

Bumps the dev-dependencies group with 8 updates in the / directory:

| Package | From | To |
| --- | --- | --- |
| [@biomejs/biome](https://github.com/biomejs/biome/tree/HEAD/packages/@biomejs/biome) | `1.8.2` | `1.8.3` |
| [@changesets/cli](https://github.com/changesets/changesets) | `2.27.6` | `2.27.7` |
| [@rspack/cli](https://github.com/web-infra-dev/rspack/tree/HEAD/packages/rspack-cli) | `0.7.4` | `0.7.5` |
| [esbu
[…]
