+++ .changeset/weak-experts-suffer.md
@@ -1,0 +1,5 @@
+---
+"create-anywidget": patch
+---
+
+Update templates to esbuild v0.23
