+++ .changeset/weak-experts-suffer.md
@@ -1,5 +1,0 @@
----
-"create-anywidget": patch
----
-
-Update templates to esbuild v0.23
