### 2024-07-24 e1d230b
refactor: static asset
