### 2024-07-24 1a1d0b8
Add trove classifiers to pyproject.toml (#619)
The framework classifiers should make anywidget discoverable within the JLab extension manager.
