### 2024-07-24 3a745d6
bring back ipywidgets
