### 2024-07-24 2d08165
feat: Patch ipywidgets to be more flexible with child widgets
