### 2024-07-28 102b688
Add marimo to readme (#631)
* Add marimo to readme

* Update README.md
