### 2024-07-30 #6939 @kylebarron
I finished my checklist! I think it's a great paper!
