### 2024-08-03 manzt/anywidget#642 @ianhi
DOC: Colab link from example does not run as is
If you click the [open colab](https://colab.research.google.com/github/manzt/anywidget/blob/main/docs/src/pages/en/notebooks/counter.ipynb) button in the counter example: https://anywidget.dev/en/notebooks/counter/

then you run into this error:

![image](https://github.com/user-attachments/assets/8768c6d6-d9d6-4fa2-877a-4c1c451d3bb8)

notably after I ran !pip install anywidget it worked subsequent times I opened it in a different tab. But for a first time user who has never run that command this could be very confusing.
