### 2024-08-03 manzt/anywidget#644 @ianhi
DOC: Confusing to transition to HMR example
While following "Getting Started" I ran into a few issues when i hit the progressive development section.

The HMR example has a few difference that make it difficult to copy paste the first example into this. I just took the inline defined counter widget and put the function into index.js and the styles in styles.css. and ran into the following issues:

1. copyable code use `index.css` which is different than the gif
2. ran into errors with `count` being undefined.
3. There is slightly different `export` placement between the inline example and HMR example in the gif
4. the `__file__` doesn't work in a notebook. (This is fine, but would be good if the example specified that we transitioned out of the notebook. I naturally was continuing this in the same notebook that i originally used for inline
