### 2024-08-03 manzt/anywidget#645 @ianhi
META: Issue templates?
Would you consider adding issue templates to this repo: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository

I find this helpful when I am opening a new issue on a repo I haven't before as it helps set expectations for what i should include, and if this is a reasonable place to do.

Examples:
https://github.com/mpl-extensions/mpl-interactions/issues/new/choose
https://github.com/matplotlib/matplotlib/issues/new/choose

You can also link out to places like your discord or github discussion for things that are questions like in the matplotlib example:

![image](https://github.com/user-attachments/assets/1ee279e1-6f63-4e0c-9c4d-d06dce038c87)
