### 2024-08-03 29892fa
chore: Add issue templates (#646)
