### 2024-08-03 manzt/anywidget#645 @manzt
Good idea!
