### 2024-08-03 manzt/anywidget#644 @manzt
Thanks for the suggestion. I'll try to update the GIF soon, but for now I've renamed the `count` to `value` (for consistency with the inline example) and provide an addendum about when `pathlib.Path` should be used. The code snippet is now copy-pasta-able in notebooks.
