### 2024-08-03 #6939 @ianhi
# Review - Comments

I worked through the examples and found the anywidget software to be easy to use and well documented. I was able to use it to quickly create a new custom widget. I found that the software works as intended and as promised makes it easier to get started developing a custom widget for use in a notebook.

I opened three issues during the review which the authors promptly addressed:

https://github.com/manzt/anywidget/issues/644
https://github.com/manzt/anywidget/issues/645
https://github.com/manzt/anywidget/issues/642

## General Checks

### Authorship

I've left this box unchecked for now.

Nezar Abdennur has no commits and Nils Gehlenborg has one commit and are both listed as authors. However there are other contributors to the software with such as tlambert03 (https://github.com/manzt/anywidget/commits?author=tlambert03) who are not listed as authors. This may be fine with the JOSS authorship policy on non-code contributions but clarification would be helpful. Policy below:

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are

## Documentation

### Functionality Documentation

> Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Yes. The docs could be made even more useful by going into slightly more detail on how to use the jupyter-widgets javascript api (e.g. `model.get`) as this may not be familiar to python developers. However, this is not API introduced by anywidget so not strictly necessary.

### Community guidelines

For support there are link to both github discussions and Discord on the github readme. These would be more prominently available if they were also placed on the documentation website.

~I also opened an issue regarding potentially adding issue templates: https://github.com/manzt/anywidget/issues/645. I think this would be nice addition but do not think it should be required.~

## Software paper

### Summary

There seems to be a discrepancy between the paper description and summary of anywidget:

> The anywidget project provides a specification and toolset for portable and reusable web-based
widgets in interactive computing environments (Figure 1). First, it defines a standard for
widget front-end code based on the web browser’s native module system. Second, it provides
tools to author, distribute, and execute these modules across web-based computing platforms.

in contrast the "What is anywidget?" section of the anywidget documentation:

> anywidget is a Python library that simplifies creating and publishing custom Jupyter Widgets. No messy build configuration or complicated cookiecutter templates.
>
> It is not a new interactive widgets framework, but rather an abstraction for creating custom Jupyter Widgets using modern web standards.

In the paper `anywidget` seems to refer both to a general system being developed for widgets which may work in a notebook, or non-notebook context, or other contexts. However the documentation seems to refer specifically to the python package that builds on top of Jupyter-widgets. This is somewhat confusing and additional clarification would be very helpful. As such I've left this box unchecked (would check it 85% percent if that was an option).

This seems to be mostly addressed in the "Custom Jupyter Widgets" section, and the confusion could be solved by using aspect of that section in the summary.

### Statement of Need

In the description of the difficulties with creating jupyter-widgets the authors write:

> However, the complexity and error-prone nature of authoring custom Jupyter Widgets have limited their adoption by the visualization community, leading authors to create or resort to alternative methods.

This is a strong statement, one that I believe this from my personal experience, but as a claim in a paper it would be useful to have concrete examples of what this difficulties are. A description of these alternative methods would be helpful as well as why anywidget is a better option.

This section would also benefit from a clear description of what the author's mean when they refer to computational notebooks. A casual reader may assume that notebooks refer specifically to Jupyter(lab) notebooks, similarly a defintion of jupyter-compatible platforms would be useful.

Integrating aspects of the "beyond jupyter" section may help with the clarity.

### State of the field

The supporting tools and ecosystem section of the paper is excellent.

The "Related Work" section also addresses this aspect, for that I have two small comments:

1. The related work section seems to omit bokeh custom widgets: https://docs.bokeh.org/en/latest/docs/user_guide/advanced/extensions.html#ug-advanced-extensions
2. Are the examples in this section are the complete set of other "alternatives" mentioned in the summary section of the paper?

## Writing Comments

Overall the writing of this paper is of high quality. There were a few sections where I believe the clarity for the reader could be enhanced with additional explanation which I have noted above in this review. Much, or perhaps all, of the necessary information for better understanding these sections is contained within the paper, however some of the necessary exposition is contained only in the final sections and the clarity could be improved with some restructuring portions of those sections into the summary or the statement of need.
