### 2024-08-16 #6939 @hugoledoux
thanks @ianhi for that thorough review.

@manzt let us know when you can address those issues. The authorship is indeed something I would like you to comment on, keeping the JOSS rules in mind: https://joss.readthedocs.io/en/latest/submitting.html#authorship

also, we are really trying to enforce papers with <1000 words, and yours at the moment is alreagy ~1600 words so while addressing the comments of @ianhi please consider this. You could simplify the text and put some parts in the docs for instance.
