### 2024-08-20 #6939 @kylebarron
> @kylebarron okay, great, thanks. But no specific comments then? Or did you comment directly to in their repo?

I think my experience might be slightly different because I've directly worked with anywidget for 12 months at this point. So I've already given @manzt a lot of feedback on the project directly.

I wasn't sure exactly what you were looking for in a JOSS review, and I felt that the project matched every checkbox, so I didn't have more to add initially.

Looking at @ianhi 's feedback, a couple things to add to the conversation:

> Yes. The docs could be made even more useful by going into slightly more detail on how to use the jupyter-widgets javascript api (e.g. `model.get`) as this may not be familiar to python developers. However, this is not API introduced by anywidget so not strictly necessary.

While agree this is useful to users, I personally consider it outside the necessary scope of anywidget. I think anywidget documents its own APIs well and discusses how it integrates with external APIs.

As a general matter, I think having some sort of API-specific documentation would be helpful. Most anywidget documentation currently is in the form of examples or prose.

> In the paper `anywidget` seems to refer both to a general system being developed for widgets which may work in a notebook, or non-notebook context, or other contexts. However the documentation seems to refer specifically to the python package that builds on top of Jupyter-widgets. This is somewhat confusing and additional clarification would be very helpful. As such I've left this box unchecked (would check it 85% percent if that was an option).

To me this is mostly a function of having two different audiences. I see the paper as trying to get at the _broader picture_ of anywidget's goals, which may be quite different from the immediate "selling point" to Python users. In general I think it's fine for the documentation website to make the Python package the primary focus.

That said, perhaps an additional page or section in the documentation that speaks to the broader ecosystem goals could be helpful.

> > However, the complexity and error-prone nature of authoring custom Jupyter Widgets have limited their adoption by the visualization community, leading authors to create or resort to alternative methods.
>
> This is a strong statement, one that I believe this from my personal experience, but as a claim in a paper it would be useful to have concrete examples of what this difficulties are. A description of these alternative methods would be helpful as well as why anywidget is a better option.

This is a fair critique. I have also had bad personal experiences with pre-anywidget Jupyter Widgets, so I accepted this as fact, but new readers who aren't familiar with the Jupyter ecosystem may not understand how painful it was to build widgets before anywidget.

> 1. The related work section seems to omit bokeh custom widgets: [docs.bokeh.org/en/latest/docs/user_guide/advanced/extensions.html#ug-advanced-extensions](https://docs.bokeh.org/en/latest/docs/user_guide/advanced/extensions.html#ug-advanced-extensions)

Other widget frameworks I've seen include [streamlit custom components](https://docs.streamlit.io/develop/concepts/custom-components/intro). I think shiny for python might have a custom non-Jupyter widget framework as well.
