### 2024-08-21 462d70a
chore: Remove examples folder (#659)
* chore: Remove examples folder

* Have ruff ignore docs
