### 2024-08-22 56ed61e
chore: Migrate from hatch to uv
