+++ .github/workflows/ci.yml
@@ -19,66 +19,50 @@ jobs:
     runs-on: macos-14
     steps:
       - uses: actions/checkout@v4
-      - uses: actions/setup-python@v5
-        with:
-          python-version: "3.12"
+      # Disable jupyter-builder build hook because not needed for linting/formatting, and requires pnpm
       - run: |
-          pip install uv
-          uv pip install --system ruff
-          ruff check
-          ruff format
+          sed -i '' 's/\[tool.hatch.build.hooks.\(.*\)\]/\[_tool.hatch.build.hooks.\1\]/' pyproject.toml
+      - name: Set up uv
+        run: curl -LsSf https://astral.sh/uv/0.3.1/install.sh | sh
+      - run: |
+          uv run ruff check
+          uv run ruff format
 
   TypecheckPython:
     name: Python / Typecheck
     runs-on: macos-14
     steps:
       - uses: actions/checkout@v4
-
-      # Disable jupyter-builder build hook because not needed for typechecking, and requires npm
+      # Disable jupyter-builder build hook because not needed for typechecking, and requires pnpm
       - run: |
           sed -i '' 's/\[tool.hatch.build.hooks.\(.*\)\]/\[_tool.hatch.build.hooks.\1\]/' pyproject.toml
-          cat pyproject.toml
-
-      - uses: actions/setup-python@v5
-        with:
-          python-version: "3.12"
-      - run: |
-          pip install uv
-          uv pip install --system -e '.[test,dev]'
-      - run: mypy
+      - name: Set up uv
+        run: curl -LsSf https://astral.sh/uv/0.3.1/install.sh | sh
+      - run: uv run mypy
 
   TestPython:
     name: Python / Test
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.7", "3.8", "3.9", "3.10", "3.11", "3.12"]
-
+        python-version:
+          - "3.8"
+          - "3.9"
+          - "3.10"
+          - "3.11"
+          - "3.12"
     steps:
       - uses: actions/checkout@v4
-
-      - uses: pnpm/action-setup@v4
-      - uses: actions/setup-node@v4
-        with:
-          node-version: "22.x"
-      - run: pnpm install --frozen-lock
-
       - uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
-
-      - if: ${{ matrix.python-version != '3.7' }}
-        run: |
-          pip install uv
-          uv pip install --system -e '.[test,dev]'
-
-      - if: ${{ matrix.python-version == '3.7' }}
-        run: |
-          pip install -e '.[test,dev]'
-
+          python-version-file: ".python-version"
+      - uses: pnpm/action-setup@v4
+        with:
+          run_install: true
+      - name: Set up uv
+        run: curl -LsSf https://astral.sh/uv/0.3.1/install.sh | sh
       - name: Run tests
-        run: pytest ./tests --color=yes --cov anywidget --cov-report xml
-
+        run: uv run --with pytest-cov pytest ./tests --color=yes --cov anywidget --cov-report xml
       - uses: codecov/codecov-action@v4
 
   LintJavaScript:
@@ -91,19 +75,15 @@ jobs:
           version: latest
       - run: biome ci .
 
-
   TypecheckJavaScript:
     name: JavaScript / Typecheck
     runs-on: macos-14
     steps:
       - uses: actions/checkout@v4
       - uses: pnpm/action-setup@v4
-      - uses: actions/setup-node@v4
         with:
-          node-version: "22.x"
-      - run: |
-          pnpm install
-          pnpm typecheck
+          run_install: true
+      - run: pnpm typecheck
 
   TestJavaScript:
[…]
