+++ pyproject.toml
@@ -11,10 +11,9 @@ authors = [
 license = { text = "MIT" }
 dynamic = ["version"]
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 dependencies =  [
     "ipywidgets>=7.6.0",
-    "importlib-metadata; python_version < '3.8'",
     "typing-extensions>=4.2.0",
     "psygnal>=0.8.1",
 ]
@@ -29,23 +28,26 @@ classifiers = [
     "Programming Language :: Python :: 3",
 ]
 
-[project.optional-dependencies]
-test = [
-    "mypy==1.11.1; python_version > '3.7'",
-    "pydantic",
-    "pytest",
-    "pytest-cov",
-    "ruff",
-    "msgspec; python_version > '3.7'",
-    "ipython<8.13; python_version < '3.9'",
-]
-dev = [
-    "comm>=0.1.0",
-    "watchfiles>=0.18.0",
-]
-
 [project.urls]
 homepage = "https://github.com/manzt/anywidget"
+
+[project.optional-dependencies]
+dev = ["watchfiles>=0.18.0"]
+
+[tool.uv]
+dev-dependencies = [
+    "comm>=0.1.4",
+    "msgspec>=0.18.6",
+    "mypy==1.11.1",
+    "pydantic>=2.5.3",
+    "pytest>=7.4.4",
+    "ruff>=0.6.1",
+    "watchfiles>=0.23.0",
+    "jupyterlab>=4.2.4",
+]
+
+[tool.uv.sources]
+anywidget = { workspace = true }
 
 [tool.hatch.build.targets.wheel.shared-data]
 "anywidget/nbextension" = "share/jupyter/nbextensions/anywidget"
@@ -53,9 +55,7 @@ homepage = "https://github.com/manzt/anywidget"
 "anywidget.json" = "etc/jupyter/nbconfig/notebook.d/anywidget.json"
 
 [tool.hatch.build]
-exclude = [
-    ".github",
-]
+exclude = [".github"]
 artifacts = [
     "anywidget/nbextension/index.*",
     "anywidget/labextension/*.tgz",
