+++ pyproject.toml
@@ -16,6 +16,7 @@ dependencies =  [
     "ipywidgets>=7.6.0",
     "typing-extensions>=4.2.0",
     "psygnal>=0.8.1",
+    "mypy>=1.11.1",
 ]
 classifiers = [
     "Framework :: Jupyter",
@@ -37,13 +38,13 @@ dev = ["watchfiles>=0.18.0"]
 [tool.uv]
 dev-dependencies = [
     "comm>=0.1.4",
-    "msgspec>=0.18.6",
-    "mypy==1.11.1",
+    "msgspec>=0.18.6; python_version > '3.7'",
     "pydantic>=2.5.3",
     "pytest>=7.4.4",
     "ruff>=0.6.1",
     "watchfiles>=0.23.0",
     "jupyterlab>=4.2.4",
+    "mypy>=1.11.1",
 ]
 
 [tool.uv.sources]
@@ -83,22 +84,6 @@ build_cmd = "build"
 [tool.hatch.version]
 path = "packages/anywidget/package.json"
 pattern = "\"version\": \"(?P<version>.+?)\""
-
-[tool.hatch.envs.default]
-features = ["test", "dev"]
-uv = true
-
-[tool.hatch.envs.default.scripts]
-lint = [
-  "ruff check {args:.}",
-  "ruff format --check --diff {args:.}",
-]
-fmt = [
-  "ruff format {args:.}",
-  "ruff check --fix {args:.}",
-]
-typecheck = "mypy anywidget"
-test = "pytest . --cov anywidget --cov-report term-missing"
 
 # https://github.com/charliermarsh/ruff
 [tool.ruff]
