### 2024-08-22 bc31487
chore: Migrate from hatch to uv (#660)
* chore: Migrate from hatch to uv

* chore: Pin python 3.11

* Fix requirements
