+++ .github/workflows/release.yml
@@ -22,28 +22,16 @@ jobs:
       issues: read
     steps:
       - uses: actions/checkout@v4
-
       - uses: pnpm/action-setup@v4
-      - uses: actions/setup-node@v4
         with:
-          node-version: "22.x"
-      - run: pnpm install --frozen-lock
-
-      - uses: actions/setup-python@v5
-        with:
-          python-version: "3.11"
-      - run: |
-          python -m pip install --upgrade pip
-          pip install build twine
-
+          run_install: true
+      - name: Set up uv
+        run: curl -LsSf https://astral.sh/uv/0.3.1/install.sh | sh
       - name: Create Release Pull Request or Publish
         id: changesets
         uses: changesets/action@v1
         with:
-          # this expects you to have a npm script called version that runs some logic and then calls `changeset version`.
-          version: npm run version
-          # This expects you to have a script called release which does a build for the packages and calls `changeset publish`
-          publish: npm run release
+          publish: pnpm release
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
