### 2024-08-27 c69750e
fix(svelte): Ensure src directory is included in npm release
