+++ .changeset/nice-jars-hug.md
@@ -1,0 +1,5 @@
+---
+"@anywidget/svelte": patch
+---
+
+Ensure all src files are included in the package release
