+++ .changeset/nice-jars-hug.md
@@ -1,5 +1,0 @@
----
-"@anywidget/svelte": patch
----
-
-Ensure all src files are included in the package release
