### 2024-08-29 a820ebe
feat(types): Make explicit widget_manager interface
