### 2024-08-29 0720abe
chore(anywidget): Make @anywidget/types a dev dependency
