### 2024-08-31 09adca8
docs: Add AFM section
