### 2024-08-31 a54e242
Update afm.md
Co-authored-by: Nezar Abdennur <nabdennur@gmail.com>
