### 2024-08-31 de058ce
Update afm.md
Co-authored-by: Nezar Abdennur <nabdennur@gmail.com>
