### 2024-08-31 6d4b6ad
Update afm.md
Co-authored-by: Nezar Abdennur <nabdennur@gmail.com>
