### 2024-08-31 b942857
Update afm.md
Co-authored-by: Nezar Abdennur <nabdennur@gmail.com>
