### 2024-08-31 5545263
Update afm.md
Co-authored-by: Nezar Abdennur <nabdennur@gmail.com>
