### 2024-08-31 3959301
docs: Add AFM section (#672)
