+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @anywidget/svelte
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`9c10efe9fd44f779fc2a5821c0e6a28ab0f4edad`](https://github.com/manzt/anywidget/commit/9c10efe9fd44f779fc2a5821c0e6a28ab0f4edad)]:
+  - @anywidget/types@0.2.0
+
 ## 0.0.9
 
 ### Patch Changes
