### 2024-08-31 3a5b386
feat(create-anywidget): Pin new @anywidget/* versions (#673)
