+++ .changeset/modern-schools-wash.md
@@ -1,0 +1,5 @@
+---
+"create-anywidget": patch
+---
+
+Upgrade framework bridges
