### 2024-08-31 41a6ceb
docs: Add anywidget narrowing Observable notebook
