### 2024-09-03 78afeb8
chore(deps-dev): bump the dev-dependencies group with 3 updates (#675)
* chore(deps-dev): bump the dev-dependencies group with 3 updates

Bumps the dev-dependencies group with 3 updates: [happy-dom](https://github.com/capricorn86/happy-dom), [@types/node](https://github.com/DefinitelyTyped/DefinitelyTyped/tree/HEAD/types/node) and [@types/react](https://github.com/DefinitelyTyped/DefinitelyTyped/tree/HEAD/types/react).


Updates `happy-dom` from 15.3.1 to 15.7.3
- [Release notes](https://github.com/capricorn86/happy-dom/releases)
- [Commits](https://github.com/capricorn86/happy-dom/compare/v15
[…]
