+++ .changeset/sweet-walls-hunt.md
@@ -1,0 +1,5 @@
+---
+"create-anywidget": patch
+---
+
+Bump types versions
