+++ .changeset/sweet-walls-hunt.md
@@ -1,5 +1,0 @@
----
-"create-anywidget": patch
----
-
-Bump types versions
