+++ .changeset/modern-schools-wash.md
@@ -1,5 +1,0 @@
----
-"create-anywidget": patch
----
-
-Upgrade framework bridges
