### 2024-09-03 bd5c313
ci: Bump uv to v0.4.3
