### 2024-09-03 8a70bb3
chore: Enable all the ruff rules
