+++ pyproject.toml
@@ -92,21 +92,7 @@ exclude = ["packages", "docs"]
 
 [tool.ruff.lint]
 pydocstyle = { convention = "numpy" }
-select = [
-    "E",    # style errors
-    "W",    # style warnings
-    "F",    # flakes
-    "D",    # pydocstyle
-    "D417", # Missing argument descriptions in Docstrings
-    "I",    # isort
-    "UP",   # pyupgrade
-    "C4",   # flake8-comprehensions
-    "B",    # flake8-bugbear
-    "A001", # flake8-builtins
-    "RUF",  # ruff-specific rules
-    "TCH",  # flake8-type-checking
-    "TID",  # flake8-tidy-imports
-]
+select = ["ALL"]
 ignore = [
     "D401", # First line should be in imperative mood (remove to opt in)
 ]
