### 2024-09-03 9c2db44
chore: Replace typing.Any with object
