+++ tests/test_experimental.py
@@ -5,7 +5,7 @@ from anywidget._descriptor import ReprMimeBundle
 from anywidget.experimental import MimeBundleDescriptor, dataclass, widget
 
 
-def test_decorator():
+def test_decorator() -> None:
     esm = "export default { render({ model , el }) {} }"
     css = ".foo { color: red;}"
 
@@ -21,7 +21,7 @@ def test_decorator():
     assert isinstance(foo._repr_mimebundle_, ReprMimeBundle)  # type: ignore
 
 
-def test_dataclass():
+def test_dataclass() -> None:
     esm = "export function render({ model , el }) {}"
     css = ".foo { color: red;}"
 
