### 2024-09-03 59a5094
chore: Fix test_widget
