+++ tests/test_widget.py
@@ -4,7 +4,7 @@ import json
 import pathlib
 import sys
 import time
-from typing import NoReturn
+from typing import Generator, NoReturn
 from unittest.mock import MagicMock, patch
 
 import anywidget
@@ -13,6 +13,7 @@ import traitlets.traitlets as t
 import watchfiles
 from anywidget._file_contents import FileContents
 from anywidget._util import _DEFAULT_ESM, _WIDGET_MIME_TYPE
+from anywidget.experimental import command
 from traitlets import traitlets
 from watchfiles import Change
 
@@ -31,7 +32,7 @@ def test_version() -> None:
 
 
 def test_basic() -> None:
-    ESM = """
+    esm = """
     function render({ model, el }) {
         el.innerText = "Hello, world";
     }
@@ -39,12 +40,12 @@ def test_basic() -> None:
     """
 
     class Widget(anywidget.AnyWidget):
-        _esm = t.Unicode(ESM).tag(sync=True)
+        _esm = t.Unicode(esm).tag(sync=True)
 
     w = Widget()
 
     assert w.has_trait("_esm")
-    assert w._esm == ESM
+    assert w._esm == esm
 
 
 def test_default_esm() -> None:
@@ -57,7 +58,7 @@ def test_default_esm() -> None:
 
 
 def test_creates_fully_qualified_identifier() -> None:
-    ESM = """
+    esm = """
     function render({ model, el }) {
         el.innerText = "Hello, world";
     }
@@ -65,7 +66,7 @@ def test_creates_fully_qualified_identifier() -> None:
     """
 
     class Widget(anywidget.AnyWidget):
-        _module = t.Unicode(ESM).tag(sync=True)
+        _module = t.Unicode(esm).tag(sync=True)
 
     w = Widget()
 
@@ -74,11 +75,11 @@ def test_creates_fully_qualified_identifier() -> None:
 
 
 def test_infer_traitlets() -> None:
-    CSS = """
+    css = """
     .foo { background-color: black; }
     """
 
-    ESM = """
+    esm = """
     function render({ model, el }) {
         el.innerText = "Hello, world";
     }
@@ -86,8 +87,8 @@ def test_infer_traitlets() -> None:
     """
 
     class Widget(anywidget.AnyWidget):
-        _esm = ESM
-        _css = CSS
+        _esm = esm
+        _css = css
 
     w = Widget()
 
@@ -99,11 +100,11 @@ def test_infer_traitlets() -> None:
 
 
 def test_infer_traitlets_partial() -> None:
-    CSS = """
+    css = """
     .foo { background-color: black; }
     """
 
-    ESM = """
+    esm = """
     function render({ model, el }) {
         el.innerText = "Hello, world";
     }
@@ -111,21 +112,21 @@ def test_infer_traitlets_partial() -> None:
     """
 
     class Widget(anywidget.AnyWidget):
-        _esm = t.Unicode(ESM).tag(foo="bar")
-        _css = CSS
+        _esm = t.Unicode(esm).tag(foo="bar")
+        _css = css
 
     w = Widget()
 
     assert w.has_trait("_esm")
-    assert w._esm == ESM
+    assert w._esm == esm
     assert w.trait_metadata("_esm", "foo") == "bar"
 
     assert w.has_trait("_css")
-    assert w._css == CSS
+    assert w._css == css
     assert w.trait_metadata("_css", "sync")
 
 
-def test_patched_repr_ipywidget_v8(monkeypatch: pytest.MonkeyPatch) -> None:
+def test_patched_repr_ipywidget_v8() -> None:
     w = anywidget.AnyWidget()
     bundle = w._repr_mimebundle_()
     assert bundle[0]
[…]
