### 2024-09-03 0c21590
Prefer type annotations
