### 2024-09-03 4e71666
Fix remaining linting issues
