+++ anywidget/__init__.py
@@ -23,7 +23,7 @@ def _jupyter_nbextension_paths() -> list[dict]:
     ]
 
 
-def load_ipython_extension(ipython) -> None:  # type: ignore[no-untyped-def]
+def load_ipython_extension(ipython) -> None:  # type: ignore[no-untyped-def]  # noqa: ANN001
     from ._cellmagic import load_ipython_extension
 
     load_ipython_extension(ipython)
