+++ pyproject.toml
@@ -95,18 +95,19 @@ pydocstyle = { convention = "numpy" }
 select = ["ALL"]
 ignore = [
     "D401", # First line should be in imperative mood (remove to opt in)
-    "COM812", # missing trailing comma (conflicts with ruff format)
-    "ISC001", # import sorting (conflicts with ruff format)
-    "DOC201", # no doc for return type (TODO: fix this)
+    "COM812", # Missing trailing comma (conflicts with ruff format)
+    "ISC001", # Import sorting (conflicts with ruff format)
+    "FIX002", # Fixable issue
+    "DOC201", # TODO(manzt) enable in follow-up PR; no doc for return type.
+    "FBT", # TODO(manzt): enable in follow-up PR; require bool options to be keyword-only.
 ]
 
 [tool.ruff.lint.per-file-ignores]
 "tests/*.py" = [
-    "D",    # no docstrings in tests
-    "S101", # use of assert
+    "D",    # No docstrings in tests
+    "S101", # Use of assert
     "B018", # "useless expression", for accessing the Foo._repr_mimbundle_ descriptor
-    "SLF001", # access private member
-    "PLC2701", # import private name
+    "SLF001", # Access private member
 ]
 "docs/*.py" = ["D"]
 
