### 2024-09-03 3b42e12
Allow use of any for get state
