### 2024-09-03 01363f9
chore: Select ALL Ruff rules (#677)
* chore: Enable all the ruff rules

* chore: Replace typing.Any with object

* chore: Apply fixes

* chore: Fix test_widget

* Prefer type annotations

* Prefer object over Any

* Fix remaining linting issues

* Allow use of any for get state
