+++ tests/test_protocols.py
@@ -1,3 +1,3 @@
-def test_protocols():
+def test_protocols() -> None:
     """Just to make sure they're imported somewhere"""
     from anywidget import _protocols  # noqa: F401
