+++ tests/test_descriptor.py
@@ -1,8 +1,9 @@
-import pathlib
+from __future__ import annotations
+
 import time
 import weakref
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, ClassVar, Set, Union
+from typing import TYPE_CHECKING, Callable, ClassVar, Generator
 from unittest.mock import MagicMock, patch
 
 import anywidget._descriptor
@@ -18,6 +19,8 @@ from anywidget._util import _WIDGET_MIME_TYPE
 from watchfiles import Change
 
 if TYPE_CHECKING:
+    import pathlib
+
     from anywidget._protocols import AnywidgetProtocol
     from ipykernel.comm import Comm
 
@@ -28,17 +31,23 @@ class MockComm(MagicMock):
 
     msg_callback = None
 
-    def on_msg(self, cb):
+    def on_msg(self, cb: Callable) -> None:
         self.msg_callback = cb
 
-    def handle_msg(self, msg):
+    def handle_msg(self, msg: object) -> None:
         if self.msg_callback is not None:
             self.msg_callback(msg)
 
 
 @pytest.fixture
-def mock_comm():
-    """Mock a comm object."""
+def mock_comm() -> Generator[MockComm, None, None]:
+    """Mock a comm object.
+
+    Yields
+    ------
+    MockComm
+        A mock comm object that can be used to simulate a comm.
+    """
     comm = MockComm()
     assert not _COMMS
     with patch.object(anywidget._descriptor, "open_comm", return_value=comm):
@@ -46,15 +55,15 @@ def mock_comm():
     assert not _COMMS
 
 
-def _send_value(comm: "Comm", value: int) -> int:
+def _send_value(comm: Comm, value: int) -> int:
     # test that the object responds to incoming messages
     comm.handle_msg(
-        {"content": {"data": {"method": "update", "state": {"value": value}}}}
+        {"content": {"data": {"method": "update", "state": {"value": value}}}},
     )
     return value
 
 
-def _assert_sends_update(wdg: "AnywidgetProtocol", comm: MagicMock, expect: int):
+def _assert_sends_update(wdg: AnywidgetProtocol, comm: MagicMock, expect: int) -> None:
     # test that the comm sends update messages
     wdg._repr_mimebundle_.send_state({"value"})
     comm.send.assert_called_with(
@@ -66,18 +75,18 @@ def _assert_sends_update(wdg: "AnywidgetProtocol", comm: MagicMock, expect: int)
 def test_descriptor(mock_comm: MagicMock) -> None:
     """Test that the descriptor decorator makes a comm, and gets/sets state."""
 
-    VAL = 1
-    REPR = "FOO"
+    val = 1
+    repr_ = "FOO"
 
     class Foo:
         _repr_mimebundle_ = MimeBundleDescriptor(autodetect_observer=False)
-        value: int = VAL
+        value: int = val
 
-        def _get_anywidget_state(self, include: Union[Set[str], None]):
+        def _get_anywidget_state(self, include: set[str] | None):  # noqa: ANN202, ARG002
             return {"value": self.value}
 
         def __repr__(self) -> str:
-            return REPR
+            return repr_
 
     foo = Foo()
     mock_comm.send.assert_not_called()  # we haven't yet created a comm object
@@ -86,11 +95,12 @@ def test_descriptor(mock_comm: MagicMock) -> None:
     mock_comm.send.assert_called_once()
     assert isinstance(repr_method, ReprMimeBundle)
     bundle = repr_method()
+    assert bundle
     assert _WIDGET_MIME_TYPE in bundle[0]  # we can call it as usual
     assert len(bundle[1]) == 0
 
     # test that the comm sends update messages
[…]
