+++ anywidget/experimental.py
@@ -14,9 +14,8 @@ if typing.TYPE_CHECKING:  # pragma: no cover
 
     from ._protocols import WidgetBase
 
-__all__ = ["dataclass", "widget", "MimeBundleDescriptor"]
+__all__ = ["MimeBundleDescriptor", "dataclass", "widget"]
 
-_T = typing.TypeVar("_T")
 T = typing.TypeVar("T")
 
 
@@ -24,7 +23,7 @@ def widget(
     *,
     esm: str | pathlib.Path,
     css: None | str | pathlib.Path = None,
-    **kwargs: typing.Any,
+    **kwargs: typing.Any,  # noqa: ANN401
 ) -> typing.Callable[[T], T]:
     """Decorator to register a widget class as a mimebundle.
 
@@ -46,7 +45,7 @@ def widget(
     if css is not None:
         kwargs["_css"] = css
 
-    def _decorator(cls: _T) -> _T:
+    def _decorator(cls: T) -> T:
         setattr(cls, "_repr_mimebundle_", MimeBundleDescriptor(**kwargs))  # noqa: B010
         return cls
 
@@ -55,13 +54,13 @@ def widget(
 
 # To preserve the signature of the decorated class.
 # see: https://github.com/pyapp-kit/magicgui/blob/5e068f31eaeeb130f43c38727b25423cc3ea4de3/src/magicgui/schema/_guiclass.py#L145-L162
-def __dataclass_transform__(
+def __dataclass_transform__(  # noqa: N807
     *,
-    eq_default: bool = True,
-    order_default: bool = False,
-    kw_only_default: bool = False,
-    field_specifiers: tuple[type | typing.Callable[..., typing.Any], ...] = (()),
-) -> typing.Callable[[_T], _T]:
+    eq_default: bool = True,  # noqa: ARG001
+    order_default: bool = False,  # noqa: ARG001
+    kw_only_default: bool = False,  # noqa: ARG001
+    field_specifiers: tuple[type | typing.Callable[..., object], ...] = (()),  # noqa: ARG001
+) -> typing.Callable[[T], T]:
     return lambda a: a
 
 
@@ -71,7 +70,7 @@ def dataclass(
     *,
     esm: str | pathlib.Path,
     css: None | str | pathlib.Path = None,
-    **dataclass_kwargs: typing.Any,
+    **dataclass_kwargs: object,
 ) -> typing.Callable[[T], T]:
     """Turns class into a dataclass, makes it evented, and registers it as a widget.
 
@@ -83,7 +82,7 @@ def dataclass(
         The path or contents of an ES Module for the widget.
     css : None | str | pathlib.Path, optional
         The path or contents of a CSS file for the widget.
-    dataclass_kwargs : typing.Any
+    dataclass_kwargs : object
         Additional keyword arguments to pass to the dataclass decorator.
 
     Returns
@@ -105,8 +104,7 @@ def dataclass(
     def _decorator(cls: T) -> T:
         cls = dataclasses.dataclass(cls, **dataclass_kwargs)  # type: ignore[call-overload]
         cls = psygnal.evented(cls)  # type: ignore[call-overload]
-        cls = widget(esm=esm, css=css)(cls)
-        return cls
+        return widget(esm=esm, css=css)(cls)
 
     return _decorator(cls) if cls is not None else _decorator  # type: ignore[return-value]
 
@@ -115,13 +113,25 @@ _ANYWIDGET_COMMAND = "_anywidget_command"
 _ANYWIDGET_COMMANDS = "_anywidget_commands"
 
 _AnyWidgetCommand = typing.Callable[
-    [typing.Any, typing.Any, typing.List[bytes]],
-    typing.Tuple[typing.Any, typing.List[bytes]],
+    [object, object, typing.List[bytes]],
[…]
