### 2024-09-06 90d0d89
ci: use setup-uv action
