+++ .github/workflows/ci.yml
@@ -22,8 +22,9 @@ jobs:
       # Disable jupyter-builder build hook because not needed for linting/formatting, and requires pnpm
       - run: |
           sed -i '' 's/\[tool.hatch.build.hooks.\(.*\)\]/\[_tool.hatch.build.hooks.\1\]/' pyproject.toml
-      - name: Set up uv
-        run: curl -LsSf https://astral.sh/uv/0.4.3/install.sh | sh
+      - uses: astral-sh/setup-uv@v2
+        with:
+          version: "0.4.6"
       - run: |
           uv run ruff check
           uv run ruff format
@@ -36,8 +37,9 @@ jobs:
       # Disable jupyter-builder build hook because not needed for typechecking, and requires pnpm
       - run: |
           sed -i '' 's/\[tool.hatch.build.hooks.\(.*\)\]/\[_tool.hatch.build.hooks.\1\]/' pyproject.toml
-      - name: Set up uv
-        run: curl -LsSf https://astral.sh/uv/0.4.3/install.sh | sh
+      - uses: astral-sh/setup-uv@v2
+        with:
+          version: "0.4.6"
       - run: uv run mypy
 
   TestPython:
@@ -59,8 +61,9 @@ jobs:
       - uses: pnpm/action-setup@v4
         with:
           run_install: true
-      - name: Set up uv
-        run: curl -LsSf https://astral.sh/uv/0.4.3/install.sh | sh
+      - uses: astral-sh/setup-uv@v2
+        with:
+          version: "0.4.6"
       - name: Run tests
         run: uv run --with pytest-cov pytest ./tests --color=yes --cov anywidget --cov-report xml
       - uses: codecov/codecov-action@v4
