### 2024-09-06 0c3b463
ci: use setup-uv action (#679)
