+++ .github/workflows/release.yml
@@ -25,8 +25,9 @@ jobs:
       - uses: pnpm/action-setup@v4
         with:
           run_install: true
-      - name: Set up uv
-        run: curl -LsSf https://astral.sh/uv/0.4.3/install.sh | sh
+      - uses: astral-sh/setup-uv@v2
+        with:
+          version: "0.4.6"
       - name: Create Release Pull Request or Publish
         id: changesets
         uses: changesets/action@v1
