+++ .changeset/thin-lies-appear.md
@@ -1,0 +1,5 @@
+---
+"create-anywidget": patch
+---
+
+Bump TypeScript to 5.6
