+++ .changeset/thin-lies-appear.md
@@ -1,5 +1,0 @@
----
-"create-anywidget": patch
----
-
-Bump TypeScript to 5.6
