+++ paper/paper.bib
@@ -461,12 +461,17 @@
   note = {Accessed: 2024-06-05}
 }
 
-@misc{jscatter,
-  author = {Fritz Lekschas},
-  title = {Jupyter Scatter},
+@article{jscatter,
+  title = {{Jupyter Scatter}: Interactive Exploration of Large-Scale Datasets},
+  author = {Fritz Lekschas and Trevor Manz},
+  journal = {Journal of Open Source Software},
+  publisher = {The Open Journal},
   year = {2024},
-  url = {https://github.com/flekschas/jupyter-scatter},
-  note = {Accessed: 2024-06-05}
+  volume = {9},
+  number = {101},
+  pages = {7059},
+  doi = {10.21105/joss.07059},
+  url = {https://doi.org/10.21105/joss.07059},
 }
 
 @ARTICLE{gos,
