+++ .github/workflows/ci.yml
@@ -22,7 +22,7 @@ jobs:
       # Disable jupyter-builder build hook because not needed for linting/formatting, and requires pnpm
       - run: |
           sed -i '' 's/\[tool.hatch.build.hooks.\(.*\)\]/\[_tool.hatch.build.hooks.\1\]/' pyproject.toml
-      - uses: astral-sh/setup-uv@v2
+      - uses: astral-sh/setup-uv@v3
         with:
           version: "0.4.6"
       - run: |
@@ -37,7 +37,7 @@ jobs:
       # Disable jupyter-builder build hook because not needed for typechecking, and requires pnpm
       - run: |
           sed -i '' 's/\[tool.hatch.build.hooks.\(.*\)\]/\[_tool.hatch.build.hooks.\1\]/' pyproject.toml
-      - uses: astral-sh/setup-uv@v2
+      - uses: astral-sh/setup-uv@v3
         with:
           version: "0.4.6"
       - run: uv run mypy
@@ -61,7 +61,7 @@ jobs:
       - uses: pnpm/action-setup@v4
         with:
           run_install: true
-      - uses: astral-sh/setup-uv@v2
+      - uses: astral-sh/setup-uv@v3
         with:
           version: "0.4.6"
       - name: Run tests
