### 2024-09-23 e07c604
docs: Clarify not just Jupyter Widgets (#691)
