+++ paper/paper.bib
@@ -675,3 +675,20 @@
   section={16.2},
   notes="Accessed 2024-06-07"
 }
+
+@misc{bokeh,
+  title = {Bokeh: Python library for interactive visualization},
+  author = {{Bokeh Development Team}},
+  year = {2018},
+  url = {https://bokeh.pydata.org/en/latest/},
+  note = {Accessed: 2024-09-23}
+}
+
+@misc{streamlit,
+  title = {Create a Component},
+  author = {{Streamlit}},
+  year = {2024},
+  url = {https://docs.streamlit.io/develop/concepts/custom-components/create},
+  note = {Accessed: 2024-09-23},
+  organization = {Snowflake Inc.}
+}
