### 2024-09-24 #6939 @manzt
Thanks for the thoughtful and detailed review, @ianhi. We have addressed your feedback both as improvements to our documentation and the manuscript.

> # Review - Comments
> I worked through the examples and found the anywidget software to be easy to use and well documented. I was able to use it to quickly create a new custom widget. I found that the software works as intended and as promised makes it easier to get started developing a custom widget for use in a notebook.
>
> I opened three issues during the review which the authors promptly addressed:
>
> manzt/anywidget#644
> manzt/anywidget#645
> manzt/anywidget#642

Thank you for the helpful suggestions. I have addressed each issue with a corresponding PR:

- https://github.com/manzt/anywidget/pull/646
- https://github.com/manzt/anywidget/pull/647
- https://github.com/manzt/anywidget/pull/643

> # General Checks
> ## Authorship
> I've left this box unchecked for now.
> Nezar Abdennur has no commits and Nils Gehlenborg has one commit and are both listed as authors. However there are other contributors to the software with such as tlambert03 (https://github.com/manzt/anywidget/commits?author=tlambert03) who are not listed as authors. This may be fine with the JOSS authorship policy on non-code contributions but clarification would be helpful. Policy below:

> > Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are

Thank you for your attention to the authorship details. Prof. Gehlenborg and Prof. Abdennur are listed as authors due to their substantial contributions beyond financial support or general supervision. They were involved in the conceptualization of the project, shaping the broader vision of “reusable widgets” for computational notebooks. Both provided active guidance throughout the project's development and contributed to drafting and editing the manuscript. Prof. Abdennur also co-developed and co-taught an anywidget tutorial at the SciPy conference.

> ## Documentation
> ### Functionality Documentation
> Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
> Yes. The docs could be made even more useful by going into slightly more detail on how to use the jupyter-widgets javascript api (e.g. model.get) as this may not be familiar to python developers. However, this is not API introduced by anywidget so not strictly necessary.

I agree that these details have been missing from the docs. I’ve addressed this by adding a new Anywidget Front-end Module (AFM) section (https://anywidget.dev/en/afm/), which includes details on the JavaScript API. This serves both widget developers and potential host platforms, outlining available APIs and implementation requirements respectively.

> ## Community guidelines
> For support there are link to both github discussions and Discord on the github readme. These would be more prominently available if they were also placed on the documentation website.

Thanks! We have updated the docs accordingly.

> ## Software paper
> ### Summary
> There seems to be a discrepancy between the paper description and summary of anywidget:
> > The anywidget project provides a specification and toolset for portable and reusable web-based
widgets in interactive computing environments (Figure 1). First, it defines a standard for
widget front-end code based on the web browser’s native module system. Second, it provides
tools to author, distribute, and execute these modules across web-based computing platforms.
>
> in contrast the "What is anywidget?" section of the anywidget documentation:
>
> > anywidget is a Python library that simplifies creating and publishing custom Jupyter Widgets. No messy build configuration or complicated cookiecutter templates. It is not a new interactive widgets framework, but rather an abstraction for creating custom Jupyter Widgets using modern web standards.
>
> In the paper anywidget seems to refer both to a general system being developed for widgets which may work in a notebook, or non-notebook context, or other contexts. However the documentation seems to refer specifically to the python package that builds on top of Jupyter-widgets. This is somewhat confusing and additional clarification would be very helpful. As such I've left this box unchecked (would check it 85% percent if that was an option).
>
> This seems to be mostly addressed in the "Custom Jupyter Widgets" section, and the confusion could be solved by using aspect of that section in the summary.

Thanks for noticing the discrepancy between the scope of anywidget and our outdated documentation. I have updated the docs to reflect the summary in our manuscript (https://github.com/manzt/anywidget/pull/691). The docs now accurately describes anywidget as both a specification and toolset for creating reusable web-based widgets in various interactive computing environments. While we've expanded the explanation to cover anywidget's broader applicability, I’ve maintained a strong focus on its primary use case of creating custom Jupyter Widgets in the documentation. I’ve also added a new Anywidget Front-end Module (AFM) section and reorganized the getting started guide to better serve users (there is a “Relationship to Jupyter Widgets” section).

> ### Statement of Need
> In the description of the difficulties with creating jupyter-widgets the authors write:
>
> > However, the complexity and error-prone nature of authoring custom Jupyter Widgets have limited their adoption by the visualization community, leading authors to create or resort to alternative methods.
>
> This is a strong statement, one that I believe this from my personal experience, but as a claim in a paper it would be useful to have concrete examples of what this difficulties are. A description of these alternative methods would be helpful as well as why anywidget is a better option.

This is a fair point. I’ve expanded on the specific difficulties of Jupyter Widget development:

- Heterogeneous Jupyter-compatible environments requiring independent targeting
- Fragmented distribution model for front-end and kernel-side code
- Cumbersome front-end development experience in notebooks

However, we would like to reiterate the broader scope of this paper, focused on the multiple related packages in the anywidget project and their overall goal. We have a [separate manuscript](https://osf.io/pyn7u/) which enumerates the challenges with Jupyter Widgets specifically. While we don't directly compare anywidget to alternatives in this section, we do discuss this comparison in the Related Work.

> This section would also benefit from a clear description of what the author's mean when they refer to computational notebooks. A casual reader may assume that notebooks refer specifically to Jupyter(lab) notebooks, similarly a defintion of jupyter-compatible platforms would be useful.

I have expanded the first paragraph to capture the generality of computational notebooks and contextualize Jupyter in this landscape.

> ## State of the field
> The supporting tools and ecosystem section of the paper is excellent.
>
> The "Related Work" section also addresses this aspect, for that I have two small comments:
The related work section seems to omit bokeh custom widgets: https://docs.bokeh.org/en/latest/docs/user_guide/advanced/extensions.html#ug-advanced-extensions

Thank you for bringing Bokeh custom widgets to our attention! We have added mention of it to the related work section.

> Are the examples in this section are the complete set of other "alternatives" mentioned in the summary section of the paper?

Besides Bokeh custom widgets, which we have now added, these are the only alternatives we are currently aware of. We have included alternative methods in the Related Work.

> ## Writing Comments
> Overall the writing of this paper is of high quality. There were a few sections where I believe the clarity for the reader could be enhanced with additional explanation which I have noted above in this review. Much, or perhaps all, of the necessary information for better understanding these sections is contained within the paper, however some of the necessary exposition is contained only in the final sections and the clarity could be improved with some restructuring portions of those sections into the summary or the statement of need.

We have done this by reorganizing the summary and statement of need to offer further clarification on computational notebooks and the relationship with Jupyter.
