### 2024-09-24 #6939 @manzt
Thanks for your thoughtful review @kylebarron.

> As a general matter, I think having some sort of API-specific documentation would be helpful. Most anywidget documentation currently is in the form of examples or prose.

We have now added the AFM section docs which describes the front-end specification detailed in this manuscript.

> To me this is mostly a function of having two different audiences. I see the paper as trying to get at the broader picture of anywidget's goals, which may be quite different from the immediate "selling point" to Python users. In general I think it's fine for the documentation website to make the Python package the primary focus.

Indeed, we've tailored our communication to different audiences. Our submission to [SciPy proceedings](https://osf.io/preprints/osf/pyn7u) addresses the Python-specific aspects (and Jupyter Widgets specifically), while this paper focuses on anywidget broader goals as a standard and general toolkit for reusable interactive widgets.

> That said, perhaps an additional page or section in the documentation that speaks to the broader ecosystem goals could be helpful.

Thanks for the suggestion! I’ve updated the docs to be consistent with this broader vision while still tailored towards the most common Jupyter Widgets use case.

> This is a fair critique. I have also had bad personal experiences with pre-anywidget Jupyter Widgets, so I accepted this as fact, but new readers who aren't familiar with the Jupyter ecosystem may not understand how painful it was to build widgets before anywidget.

We have addressed these specific challenges in our SciPy manuscript, which focuses on the Python library and enumerates the challenges with authoring Jupyter Widgets. We now reference this work in this manuscript.

> Other widget frameworks I've seen include streamlit custom components. I think shiny for python might have a custom non-Jupyter widget framework as well.

Thanks for these suggestions. I’ve now added them to the related work!
