### 2024-09-24 #6939 @manzt
@hugoledoux We've addressed @ianhi & @kylebarron's comments primarily in expanded documentation, minimizing manuscript additions. Although the manuscript exceeds JOSS's desired length, we believe the current scope is necessary to adequately convey the breadth of the anywidget project and ecosystem.

Thanks for your feedback and patience, everyone!
