### 2024-09-24 9dfa2ba
ci: Setup build/publish with uv
