+++ .github/workflows/ci.yml
@@ -24,7 +24,7 @@ jobs:
           sed -i '' 's/\[tool.hatch.build.hooks.\(.*\)\]/\[_tool.hatch.build.hooks.\1\]/' pyproject.toml
       - uses: astral-sh/setup-uv@v2
         with:
-          version: "0.4.6"
+          version: "0.4.16"
       - run: |
           uv run ruff check
           uv run ruff format
@@ -39,7 +39,7 @@ jobs:
           sed -i '' 's/\[tool.hatch.build.hooks.\(.*\)\]/\[_tool.hatch.build.hooks.\1\]/' pyproject.toml
       - uses: astral-sh/setup-uv@v2
         with:
-          version: "0.4.6"
+          version: "0.4.16"
       - run: uv run mypy
 
   TestPython:
@@ -63,7 +63,7 @@ jobs:
           run_install: true
       - uses: astral-sh/setup-uv@v2
         with:
-          version: "0.4.6"
+          version: "0.4.16"
       - name: Run tests
         run: uv run --with pytest-cov pytest ./tests --color=yes --cov anywidget --cov-report xml
       - uses: codecov/codecov-action@v4
