+++ .github/workflows/release.yml
@@ -27,7 +27,7 @@ jobs:
           run_install: true
       - uses: astral-sh/setup-uv@v2
         with:
-          version: "0.4.6"
+          version: "0.4.16"
       - name: Create Release Pull Request or Publish
         id: changesets
         uses: changesets/action@v1
@@ -36,5 +36,5 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.TWINE_API_KEY }}
+          UV_PUBLISH_USERNAME: __token__
+          UV_PUBLISH_PASSWORD: ${{ secrets.TWINE_API_KEY }}
