+++ .github/workflows/release.yml
@@ -25,7 +25,7 @@ jobs:
       - uses: pnpm/action-setup@v4
         with:
           run_install: true
-      - uses: astral-sh/setup-uv@v2
+      - uses: astral-sh/setup-uv@v3
         with:
           version: "0.4.16"
       - name: Create Release Pull Request or Publish
