### 2024-09-24 94bc169
ci: Setup build/publish with uv (#692)
* ci: Setup build/publish with uv

* Upgrade to v3
