+++ .changeset/clever-lobsters-type.md
@@ -1,0 +1,5 @@
+---
+"create-anywidget": patch
+---
+
+Use esbuild v0.24.0
