+++ .changeset/clever-lobsters-type.md
@@ -1,5 +1,0 @@
----
-"create-anywidget": patch
----
-
-Use esbuild v0.24.0
