### 2024-09-26 #6939 @hugoledoux
@ianhi when you have some time, could you check that the updates are satisfactory to you? And click the checkboxes above? It seems to me they are, but I'd like you to confirm.
