### 2024-09-26 902c867
ci: Revert back to twine for uploading to CI
