+++ pyproject.toml
@@ -24,9 +24,10 @@ classifiers = [
 ]
 dependencies =  [
     "comm>=0.1.4",
-    "importlib-metadata; python_version < '3.8'",
-    "typing-extensions>=4.2.0",
+    "jupyterlab-widgets>=3.0.13",
     "psygnal>=0.8.1",
+    "traitlets>=5.14.3",
+    "widgetsnbextension>=4.0.13",
 ]
 
 [project.optional-dependencies]
@@ -43,6 +44,7 @@ dev-dependencies = [
     "pydantic>=2.5.3",
     "pytest>=7.4.4",
     "ruff>=0.6.1",
+    "typing-extensions>=4.12.2",
     "watchfiles>=0.23.0",
 ]
 
