### 2024-10-17 #6939 @ianhi
Very nice! The paper reads well and is very informative, all issues I had have been resolved so I've checked all the boxes.

Anywidget is really a great project Trevor, I'm glad we have it.
