### 2024-10-21 #6939 @manzt
>  harmonise "front-end" with the hyphen everywhere: manzt/anywidget@9e71a71/paper/paper.md?plain=1#L64C13-L64C20

People seem to have differing opinions about frontend vs front-end vs front end. I've used the term "front-end" when it is used as a compound adjective and "front end" when used as a noun (a recommendation I've seen [several places](https://css-tricks.com/poll-results-front-end-front-end/#aa-the-right-answer-seems-to-be)). If preferred we can switch to front-end everywhere, but that was the intension.
