### 2024-10-21 #6939 @hugoledoux
You obviously have put more thoughts into this than me (never thought of that!), so we do it your way.
