+++ paper/paper.bib
@@ -10,8 +10,7 @@
 }
 
 @misc{JupyterBookMyst,
-  author       = {Holdgraf, Chris and
-                  The Executable Books Community},
+  author       = {Holdgraf, Chris and others},
   title        = {{Jupyter Book and MyST: A community-led, 
                    extensible, modular ecosystem for creating
                    computational narratives}},
