+++ paper/paper.md
@@ -229,8 +229,12 @@ documentation about the project can be found at https://anywidget.dev.
 Interactive notebook visualization tools vary widely in features and
 compatibility [@Wang2024]. Some tools offer rich features (e.g., bi-directional
 communication) but rely on platform-specific APIs, limiting compatibility
-[@Zhao2022; @Wang2023; @Drosos2020; @Li2023; @Jain2022]. More simple approaches
-provide broader compatibility but lack features which meaningfully enrich user
+[@Zhao2022; @Wang2023; @Drosos2020; @Li2023; @Jain2022]. For example,
+frameworks like [Bokeh](https://bokeh.pydata.org/en/latest) and
+[Streamlit](https://docs.streamlit.io/develop/concepts/custom-components/create)
+enable custom extensions, but these integrations are tied to their specific
+frameworks and cannot be reused elsewhere. More simple approaches provide
+broader compatibility but lack features which meaningfully enrich user
 workflows. For example, using static templates or the NOVA framework
 [@Wang2022] offers wide compatibility, as the resulting HTML displays can be
 embedded in nearly any web-based notebook platform. However, this approach
