### 2024-10-22 #6939 @manzt
> First I proof-read the paper, and if you could fix those:
- [x] harmonise "front-end" with the hyphen everywhere: https://github.com/manzt/anywidget/blob/9e71a71ae6daaf9f892ccba21ed3762d118dded5/paper/paper.md?plain=1#L64C13-L64C20
- [x] some of your references where there are no "human" authors end up being weird references because of the way latex parses the names: (Project, 2024), (Community, 2022), (Streamlit 2024), and (QuantStack, 2019). I suggest just using Markdown links for those and link to the URL

> Then, could you:
