### 2024-10-22 059084b
Update JOSS Paper (#705)
* chore: Add missing DOIs

* chore: Remove abstract

* Add Acknowledgements section

* chore: Update jscatter ref

* Add Funding

* Add additional related work

* Update statement of need

* Update manuscript

* Use hrefs
