+++ paper/paper.md
@@ -1,15 +1,5 @@
 ---
 title: 'anywidget: reusable widgets for interactive analysis and visualization in computational notebooks'
-abstract: |
-  The anywidget project provides a specification and toolset for creating
-  portable web-based widgets for computational notebooks. It defines a standard
-  for front-end widget code using ECMAScript modules and offers tools for
-  authoring, distributing, and executing these modules across Jupyter-compatible
-  platforms (e.g., JupyterLab, Google Colab, VS Code). The project has spurred
-  the creation of many interactive visualization libraries and has been
-  integrated into existing ones. The standard has been embraced by other
-  notebook platforms and web frameworks, extending the reach of interactive
-  widgets beyond the Jupyter ecosystem.
 tags:
   - Computational notebooks
   - Jupyter
@@ -54,11 +44,11 @@ reusable web-based widgets in interactive computing environments
 code based on the web browser's native module system. Second, it provides tools
 to author, distribute, and execute these modules across web-based computing
 platforms. Since its release a year and a half ago, anywidget has steadily
-gained adoption. Nearly 70 new widgets have been created or ported to anywidget
-and published to the Python Package Index (PyPI), along with many standalone
-scripts and notebooks. These tools cover general-purpose visualization
-libraries [@jscatter; @Heer2024] as well as notebook integrations for
-applications in biology [@gos; @vitessce; @viv; @cev], mapping [@lonboard],
+gained adoption. Nearly 100 new widgets have been created or ported to
+anywidget and published to the Python Package Index (PyPI), along with many
+standalone scripts and notebooks. These tools cover general-purpose
+visualization libraries [@jscatter; @Heer2024] as well as notebook integrations
+for applications in biology [@gos; @vitessce; @viv; @cev], mapping [@lonboard],
 astronomy [@ipyaladin], and education [@drawdata]. Anywidget has also been
 integrated into popular visualization libraries like Altair [@altair],
 enhancing interactivity in notebooks and deepening user engagement with
@@ -69,27 +59,32 @@ visualizations and code.
 # Statement of need
 
 Computational notebooks are the preferred environment for interactive computing
-and data analysis. Their popularity has spurred the development of interactive
-visual analytics systems that integrate seamlessly within these environments
-[@Wang2024]. The Jupyter project [@Kluyver2016; @Perez2007; @Granger2021] has
-fostered an ecosystem for writing, executing, and sharing computational
-notebooks, including tools for converting them into formats like
-books [@JupyterBook; @JupyterBookMyst], presentation slides [@nbconvert;
-@Wang2023], and dashboards [@voila]. However, approaches for authoring
-interactive notebook visualizations vary widely in features and platform
-compatibility [@Wang2024], resulting in diverse yet incompatible systems.
-This inconsistency hinders composition and reuse of interactive visualizations
-in notebooks, fostering platform-specific, monolithic integrations instead of
-reusable, modular components that fit various analysis workflows.
+and data analysis. These platforms provide an interface where users can write,
+execute, and interact with code and data in real-time. Notebooks connect an
+interactive front end (typically a web browser) to an external system that runs
+the code, combining prose, executable snippets, and media. The widespread
+adoption of notebooks has driven the development of tools for integrating
+interactive visualizations within these environments [@Wang2024]. The Jupyter
+project [@Kluyver2016; @Perez2007; @Granger2021] has become synonymous with
+computational notebooks, largely due to its widespread success in fostering an
+extensive ecosystem of community tools (e.g., for producing books
+[@JupyterBookMyst], presentation slides [@Wang2023], and dashboards (e.g.,
+[Voilà](https://voila.readthedocs.io))).
 
-No authoring approach aligns interactive visualization systems with the broader
-notebook ecosystem, except for Jupyter Widgets, which show potential due to
[…]
