### 2024-10-23 #6939 @hugoledoux
I am curious: why is the version not just "0.7.3", why append "create-anywidget@" to the version?

I guess this won't break JOSS archiving system, but I am first and foremost curious.
