### 2024-10-23 #6939 @manzt
Good question! anywidget is a monorepo with several different [packages]( https://github.com/manzt/anywidget?tab=readme-ov-file#packages). Each package is versioned independently and a releases are made automatically by CI using a tag prefixed with the package name whenever that package changes. We use a tool called changesets to manage all of this.

“create-anywidget” is a cli we we publish to npm for bootstrapping new “anywidget” Python projects. It has separate version and releases.
