### 2024-10-23 #6939 @manzt
I linked the “anywidget@0.9.13” release in the comment as well as that is the latest release of the primary python package most folks interface with. However “create-anywidget@0.7.3” is the latest tag for the entire project. I suppose we could have some monotonic version number for the entire suite of package (ie the anywidget project) but it wouldn’t be semantically meaningful and would just add noise to the releases/changeligs.
