### 2024-10-23 #6939 @hugoledoux
@editorial check references
