### 2024-10-23 #6939 @hugoledoux
@manzt all the errors come from the fact that the branch the bot uses was still "joss" but you had removed it, I put "main"
