### 2024-10-23 #6939 @hugoledoux
thank you @kylebarron @ianhi for your reviews and feedback. I hope you enjoyed it and learned something in the process!

@manzt, one editor-in-chief of JOSS, will now pick it up and formally accept it. It's possible that they request a few minor changes. This should happen within 1-2 days.
