### 2024-10-23 9b809d8
minor changes in paper.md
