### 2024-10-23 #6939 @danielskatz
@manzt - I've suggested some small changes in https://github.com/manzt/anywidget/pull/706 - please merge this or let me know what disagree with, then we can proceed
