### 2024-10-23 b01035b
Remove myst config
