### 2024-10-23 cf258a0
chore: Apply Nezar nits to manuscript (#707)
* chore: Apply Nezar nits

* Remove myst config
