### 2024-10-23 3a6063c
minor changes for JOSS publication (#706)
* minor changes in paper.md

* minor changes in paper.bib
