+++ paper/paper.bib
@@ -114,7 +114,7 @@
 
 @INPROCEEDINGS{Wang2023,
   title     = "{Slide4N}: Creating Presentation Slides from Computational
-               Notebooks with {Human-AI} Collaboration",
+               Notebooks with Human-{AI} Collaboration",
   booktitle = "Proceedings of the 2023 {CHI} Conference on Human Factors in
                Computing Systems",
   author    = "Wang, Fengjie and Liu, Xuye and Liu, Oujing and Neshati, Ali and
@@ -151,7 +151,7 @@
 }
 
 @INPROCEEDINGS{Drosos2020,
-  title     = "Wrex: A Unified {Programming-by-Example} Interaction for
+  title     = "Wrex: A Unified Programming-by-Example Interaction for
                Synthesizing Readable Code for Data Scientists",
   booktitle = "Proceedings of the 2020 {CHI} Conference on Human Factors in
                Computing Systems",
@@ -259,7 +259,7 @@
 }
 
 @ARTICLE{Wang2022,
-  title         = "{NOVA}: A Practical Method for Creating {Notebook-Ready}
+  title         = "{NOVA}: A Practical Method for Creating Notebook-Ready
                    Visual Analytics",
   author        = "Wang, Zijie J and Munechika, David and Lee, Seongmin and
                    Chau, Duen Horng",
@@ -548,7 +548,7 @@
 
 @INPROCEEDINGS{ipyaladin,
        author = {{Boch}, Thomas and {Desroziers}, J{\'e}r{\^o}me},
-        title = "{ipyaladin: Enabling Aladin Lite in Jupyter Notebooks}",
+        title = {{ipyaladin}: Enabling {A}ladin {L}ite in {J}upyter Notebooks},
     booktitle = {Astronomical Data Analysis Software and Systems XXVII},
          year = 2020,
        editor = {{Ballester}, Pascal and {Ibsen}, Jorge and {Solar}, Mauricio and {Shortridge}, Keith},
@@ -575,7 +575,7 @@
 }
 
 @ARTICLE{altair,
-  title     = "Altair: Interactive statistical visualizations for python",
+  title     = "Altair: Interactive statistical visualizations for {P}ython",
   author    = "VanderPlas, Jacob and Granger, Brian and Heer, Jeffrey and
                Moritz, Dominik and Wongsuphasawat, Kanit and Satyanarayan,
                Arvind and Lees, Eitan and Timofeev, Ilia and Welsh, Ben and
@@ -602,7 +602,7 @@
 
 @misc{ipywidgets,
   author = {Jason Grout and Jonathan Frederic and Sylvain Corlay and et al.},
-  title = {ipywidgets: Interactive widgets for the Jupyter Notebook},
+  title = {{ipywidgets}: Interactive widgets for the {J}upyter Notebook},
   year = {2024},
   url = {https://github.com/jupyter-widgets/ipywidgets},
   note = {Accessed: 2024-06-05}
