### 2024-10-23 00d734e
docs: Drop Jupyter in packages section
