### 2024-10-23 #6939 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
