### 2024-10-23 #6939 @danielskatz
👋 @manzt - please check this PDF and confirm that it looks good to you, then I'll accept and publish it.
