### 2025-02-06 #6939 @manzt
@danielskatz this manuscript was published last year but hasn't seem to be picked up on Google Scholar. I found some old thread where it seemed like an erroneous `\n` in the title in the CrossRef metadata may cause some downstream issues. I checked the CrossRef metadata for this paper (https://api.crossref.org/works/10.21105/joss.06939) and it seems there is also a `\n` in the title. Is it possible to follow a procedure similar to ensure the erroneous `\n` is excluded?

- https://github.com/openjournals/joss/issues/1376#issuecomment-2519714716
- https://github.com/openjournals/joss-reviews/issues/6889#issuecomment-2519661513

I'm not sure if this would resolve the problem with indexing with Google Scholar, but I figured it was worth asking for a correction.
