### 2025-02-06 #6939 @danielskatz
What happens inside Google Scholar is a mystery to all...  But we can try this.
