### 2025-03-14 #6939 @manzt
> Let me reaccept the paper, which will rebuild it, then we can see what happens

Looks like it showed up! No idea how, but thank you for helping! https://scholar.google.com/citations?user=Lo7nJd0AAAAJ&hl=en
